🔄 Refactor : clean-up structure
- classify with its functionality
