+++ lib/engine/.gitignore
@@ -1,1 +1,3 @@
-code
+out
+../../tests/code
+foundry
