+++ lib/etherscan/__init__.py
@@ -1,5 +1,10 @@
 __all__ = ["unichain", "verify_script"]
 
+import os
+
 from dotenv import load_dotenv
 
 load_dotenv()
+_network = "unichain"
+_cache_base = os.path.join("code", _network)
+foundry_dir: str = os.path.join("code", "unichain")
