+++ lib/extractor/__init__.py
@@ -1,0 +1,5 @@
+__all__ = ["getter"]
+
+import dotenv
+
+dotenv.load_dotenv()
