--- lib/parser/.gitignore
@@ -1,3 +1,0 @@
-out
-../../tests/code
-foundry
