--- lib/parser/__init__.py
@@ -1,7 +1,0 @@
-__all__ = ["run_semgrep", "layer_2", ]
-
-import os
-
-from utils import project_root_abs
-
-os.chdir(project_root_abs)
