+++ tests/test_eval.py
@@ -1,71 +1,4 @@
-class Logic:
-    def __init__(self, native: str):
-        self.native = native
-
-    def __str__(self):
-        return self.native
-
-    def __repr__(self):
-        return self.native
-
-    def __eq__(self, other):
-        return str(self) == str(other)
-
-    def __or__(self, other: "Logic"):
-        if self == "" or other == "":
-            return Logic(f"{self}{other}")
-        return Logic(f"(({self}) or ({other}))")
-
-    def __and__(self, other: "Logic"):
-        if self == "" or other == "":
-            return Logic(f"{self}{other}")
-        return Logic(f"(({self}) and ({other}))")
-
-    def __invert__(self):
-        if self == "" or self is None:
-            return self
-        return Logic(f"not({self})")
-
-
-class LogicNode:
-    def __init__(self, logic: "Logic", parent: "LogicNode" = None):
-        self.logic = logic
-        self._parent: LogicNode = self if parent is None else parent
-        # if parent is not None:
-        #     print("parent", parent, "cur", self)
-        self.children: list[LogicNode] = []
-
-    @property
-    def parent(self):
-        return self._parent
-
-    def __str__(self):
-        return self.logic.__str__()
-
-    @property
-    def peers(self):
-        return self.parent.children
-
-    def concat_previous_children(self, condition: str) -> Logic:
-        _children = self.children
-        if len(_children) == 0:
-            return Logic(condition)
-        if condition:  # else if
-            return Logic(condition) & ~_children[-1].logic
-        return ~_children[-1].logic  # else
-
-
-def add_child(parent: LogicNode, condition: str = "") -> LogicNode:
-    child_node = LogicNode(
-        parent.logic &
-        parent.concat_previous_children(condition)
-        , parent)
-    # print("child", child_node)
-    parent.children.append(child_node)
-    return child_node
-
-
-_code_root_node = LogicNode(Logic(""), None)  # assume there is an empty root
+from engine.lexer import Logic, LogicNode, add_child
 
 
 def test_logic_operations():
